+++ src/components/core/layout/Content/Content-styles.ts
@@ -1,3 +1,5 @@
 import styled from 'styled-components';
 
-export const Content = styled('div')({});
+export const Content = styled('div')<{ width?: string }>(({ width }) => ({
+  width,
+}));