+++ src/components/core/layout/Content/Content-types.ts
@@ -1,4 +1,5 @@
 export interface ViewProps {
   /** - */
   children?: React.ReactNode;
+  width?: string;
 }