+++ src/components/core/layout/Content/Content-view.tsx
@@ -3,8 +3,10 @@ import { ViewProps } from './Content-types';
 import { Content } from './Content-styles';
 
 /** A semantic wrapper component without special meaning for grouping content. */
-export const ContentView: React.FC<ViewProps> = React.memo(({ children }) => (
-  <Content className="Content">{children}</Content>
+export const ContentView: React.FC<ViewProps> = React.memo(({ width, children }) => (
+  <Content className="Content" width={width}>
+    {children}
+  </Content>
 ));
 
 export default ContentView;